# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BougieBams API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Unauthorized
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  # ── Events ────────────────────────────────────────────────────────────────
...
  /subscribe:
    post:
      operationId: subscribe
      tags: [subscribers]
      summary: Subscribe to the mailing list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResult"
        "409":
          description: Already subscribed
  /admin/subscribers:
    get:
      operationId: listSubscribers
      tags: [subscribers]
      summary: List all subscribers (admin)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Subscribers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberListResponse"
        "401":
          description: Unauthorized

  # ── Contact ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiProduct:
      type: object
      required: [id, sku, name, description, price, category, inStock]
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean
        imagePath:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [sku, name, description, price, category, inStock]
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        inStock:
          type: boolean
    ProductListResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ApiProduct"
    ProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/ApiProduct"

    # ── Event schemas ────────────────────────────────────────────────────
...
    SubscribeInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        source:
          type: string
        discountCode:
          type: string
    SubscribeResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
        discountCode:
          type: ["string", "null"]
    Subscriber:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        source:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        createdAt:
          type: string
    SubscriberListResponse:
      type: object
      required: [subscribers]
      properties:
        subscribers:
          type: array
          items:
            $ref: "#/components/schemas/Subscriber"

    # ── Contact schemas ──────────────────────────────────────────────────
...