# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  sku: string;
  name: string;
  description: string;
  price: number;
  category: string;
  inStock: boolean;
}