# lib/api-zod/src/generated/types/productResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiProduct } from './apiProduct';

export interface ProductResponse {
  product: ApiProduct;
}