# lib/api-zod/src/generated/types/subscribeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeInput {
  email: string;
  source?: string;
  discountCode?: string;
}