# lib/api-zod/src/generated/types/subscribeResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscribeResult {
  message: string;
  /** @nullable */
  discountCode?: string | null;
}