# lib/api-zod/src/generated/types/subscriberListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * BougieBams API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Subscriber } from './subscriber';

export interface SubscriberListResponse {
  subscribers: Subscriber[];
}